+++ Carwash/src/LoadingBar.h
@@ -3,21 +3,21 @@
 
 class LoadingBar {
     private:
-        uint8_t start;
-        uint8_t currentValue;
-        uint8_t end;
-        uint8_t currentPercentage;
-        uint8_t blockPercentage;
-        uint8_t barSize;
+        unsigned long start;
+        unsigned long currentValue;
+        unsigned long end;
+        unsigned long currentPercentage;
+        unsigned long blockPercentage;
+        unsigned long barSize;
         bool changed = true;
         String cache;
-        uint8_t calculatePercentage(uint8_t start, uint8_t value, uint8_t end);
+        unsigned long calculatePercentage(unsigned long start, unsigned long value, unsigned long end);
     public:
-        LoadingBar(uint8_t start, uint8_t currentValue, uint8_t end,
-            uint8_t barSize);
-        void setCurrentValue(uint8_t currentValue);
-        void forcePercentage(uint8_t percentage);
-        uint8_t getPercentage();
+        LoadingBar(unsigned long start, unsigned long currentValue, unsigned long end,
+            unsigned long barSize);
+        void setCurrentValue(unsigned long currentValue);
+        void forcePercentage(unsigned long percentage);
+        unsigned long getPercentage();
         String getLoadingBar();
 };
 #endif
